feat: Implement LinksGroup component for improved navigation in dashboard
